Mandatory_01/01_data_formats/python/pyproject.toml
[tool.poetry]
name = "01_data_formats"
version = "0.1.0"
description = ""
authors = ["MikkelYtting <62931295+MikkelYtting@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
